tweak(bot): links

+++ arma-bot/bot.js
@@ -200,7 +200,7 @@ client.getPerms = function(msg) {
 
 client.on('message', (message) => {
     if (!message.author.bot){
-        if (message.content.includes('discord.gg/')){
+        if (message.content.includes('discord.gg/') && !message.author.id === "609044650019258407"){
             if (!message.content.includes('discord.gg/armarp')){
                 message.delete()
                 return
@@ -246,7 +246,7 @@ client.on('message', (message) => {
                     msg.delete(5000)
                 })
                 return
-            }else  if (!message.channel.name.includes('bot') && !message.channel.name.includes('verify')) {
+            }else if (!message.channel.name.includes('bot') && !message.channel.name.includes('verify')) {
                 message.delete()
                 message.reply('Please use bot commands for this command.').then(msg => {
                     msg.delete(5000)
